+++ index.js
@@ -58,6 +58,17 @@ function requireAllModels(Implementation, modelsDir, displayMessage) {
   }
 }
 
+function getFields(opts) {
+  return _.map(opts.fields, function (field) {
+    field.isVirtual = true;
+    field.isSearchable = false;
+    field.isSortable = field.isSortable || false;
+    field.isReadOnly = !field.set;
+
+    return field;
+  });
+}
+
 exports.Schemas = Schemas;
 exports.logger = logger;
 exports.ResourcesRoute = {};
@@ -208,7 +219,7 @@ exports.init = function (Implementation) {
               attributes: ['field', 'displayName', 'type', 'enums',
                 'collection_name', 'reference', 'column', 'isSearchable',
                 'widget', 'integration', 'isReadOnly', 'isVirtual',
-                'isRequired', 'defaultValue', 'validations']
+                'isRequired', 'defaultValue', 'validations', 'isSortable']
             },
             validations: {
               attributes: ['type', 'value', 'message']
@@ -305,15 +316,8 @@ exports.collection = function (name, opts) {
     Schemas.schemas[name].actions = _.union(opts.actions, Schemas.schemas[name].actions);
     Schemas.schemas[name].segments = _.union(opts.segments, Schemas.schemas[name].segments);
 
-    opts.fields = _.map(opts.fields, function (field) {
-      // NOTICE: Smart Field definition case
-      field.isVirtual = true;
-      field.isSearchable = false;
-      field.isReadOnly = !field.set;
-
-      return field;
-    });
-
+    // NOTICE: Smart Field definition case
+    opts.fields = getFields(opts);
     Schemas.schemas[name].fields = _.concat(opts.fields,
       Schemas.schemas[name].fields);
 
@@ -325,6 +329,7 @@ exports.collection = function (name, opts) {
     opts.name = name;
     opts.isVirtual = true;
     opts.isSearchable = !!opts.isSearchable;
+    opts.fields = getFields(opts);
     Schemas.schemas[name] = opts;
   }
 };