+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Smart Fields - Add a parameter to specify if the sorting is allowed on this field.
+
 ### Fixed
 - Initialization - Ignore directories while loading models.